refactor: move tool_definition.py to pydantic_ai_integration
ARCHITECTURE IMPROVEMENT:
- Moved tool_definition.py from pydantic_models/tool_session/ to pydantic_ai_integration/
- Better separation of concerns: tool infrastructure now lives together
- pydantic_models remains pure data structures (ToolRequest, ToolResponse, ToolEvent)
- pydantic_ai_integration contains tool management logic (registration, validation, execution)

CHANGES:
- src/pydantic_models/tool_session/tool_definition.py → src/pydantic_ai_integration/tool_definition.py
- Updated import in tool_decorator.py to use relative import (.tool_definition)
- No functional changes, purely organizational

RATIONALE:
- ManagedToolDefinition contains business logic (validate_params, check_permission, check_enabled)
- Stores executable references (implementation function, params_model class)
- Conceptually part of tool registration/execution infrastructure
- Decorator and definition should be co-located

Resume endpoint cleanup also included:
- Removed resume_models.py (obsolete)
- Removed /resume endpoint from router
- Removed resume_session() from service
- Resume is now implicit via session_id
